requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
opencv-python-headless==4.8.0.74
pillow==9.5.0
numpy==1.24.3
gunicorn==21.2.0
requests==2.31.0
psutil==5.9.5
gradio_client==1.12.1